+++ cmd2/constants.py
@@ -41,6 +41,9 @@ CMD_ATTR_ARGPARSER = 'argparser'
 # Whether or not tokens are unquoted before sending to argparse
 CMD_ATTR_PRESERVE_QUOTES = 'preserve_quotes'
 
+# Deprecation metadata (a cmd2.CommandDeprecation) for a command marked with @with_deprecation
+CMD_ATTR_DEPRECATED = 'deprecation'
+
 # subcommand attributes for the base command name and the subcommand name
 SUBCMD_ATTR_COMMAND = 'parent_command'
 SUBCMD_ATTR_NAME = 'subcommand_name'